Update helpers and add PasswordTest
