+++ src/org/labkey/test/tests/UserTest.java
@@ -592,6 +592,23 @@ public class UserTest extends BaseWebDriverTest
         assertElementPresent(Locators.labkeyErrorSubHeading.containing("This request has an invalid security context. You may have signed in or signed out of this session. Try again by using the 'back' and 'refresh' button in your browser."));
     }
 
+    @Test
+    public void testDisplayName()
+    {
+        final UIUserHelper uiUserHelper = new UIUserHelper(this);
+
+        String newDisplayName = "changeDisplayTest";
+
+        uiUserHelper.setDisplayName(NORMAL_USER, newDisplayName);
+        assertTextPresent(newDisplayName);
+
+        String injectDisplayName = "displayName" + INJECT_CHARS_1;
+
+        uiUserHelper.setDisplayName(NORMAL_USER, injectDisplayName);
+        assertTextPresent(injectDisplayName);
+        assertTextNotPresent(newDisplayName);
+    }
+
     @Override
     public BrowserType bestBrowser()
     {
